systems/quest/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiomysql==0.2.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
blinker==1.9.0
blis==0.7.11
bs4==0.0.2
catalogue==2.0.10
certifi==2022.12.7
cffi==1.17.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.8
cloudpathlib==0.21.1
confection==0.1.5
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
debugpy==1.8.14
demjson3==3.0.6
distro==1.9.0
duckdb==1.3.2
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl#sha256=6a0f857a2b4d219c6fa17d455f82430b365bf53171a2d919b9376e5dc9be032e
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et_xmlfile==2.0.0
exceptiongroup==1.2.2
faiss-cpu==1.10.0
filelock==3.13.1
Flask==3.1.0
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.6.1
gdown
google-generativeai
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==2.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-experimental==0.3.4
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
langcodes==3.5.0
langsmith==0.3.28
language_data==1.3.0
litellm==1.65.4.post1
loguru==0.5.3
lotus-ai==1.1.2
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.1
mdurl==0.1.2
modelscope==1.27.1
mpmath==1.3.0
multidict==6.4.2
murmurhash==1.0.13
mypy-extensions==1.0.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
ollama==0.4.8
openai==1.76.0
opengauss-sqlalchemy==2.4.0
openpyxl==3.1.5
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pdfminer.six==20250506
pdfplumber==0.11.7
pgvector==0.4.1
pillow==11.0.0
ply==3.11
preshed==3.0.10
propcache==0.3.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pydot==4.0.1
Pygments==2.19.2
PyMySQL==1.1.1
pyobvector==0.2.11
pyparsing==3.2.3
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
sqlglot==26.15.0
sql_metadata
srsly==2.5.1
sympy==1.13.1
tenacity==9.1.2
thinc==8.2.5
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.19.1
torch==2.1.2+cu121
torchaudio==2.1.2+cu121
torchvision==0.16.2+cu121
tqdm==4.67.1
transformers==4.41.0
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==1.26.13
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wordninja==2.0.0
wrapt==1.17.2
yarl==1.19.0
zipp==3.21.0
zstandard==0.23.0